Refactoring constants.
While working on the Jira issue https://4teamwork.atlassian.net/browse/GEVER-122, my mentor asked me to refactor one of the constants used for the database schemata.

The reason for this change is to have all constants at the same place.
