+++ opengever/globalindex/model/__init__.py
@@ -1,5 +1,3 @@
-WORKFLOW_STATE_LENGTH = 255
-
 tables = [
     'task_principals',
     'tasks',
