+++ opengever/meeting/model/proposal.py
@@ -2,6 +2,7 @@ from opengever.base import advancedjson
 from opengever.base.model import Base
 from opengever.base.model import UNIT_ID_LENGTH
 from opengever.base.model import USER_ID_LENGTH
+from opengever.base.model import WORKFLOW_STATE_LENGTH
 from opengever.base.oguid import Oguid
 from opengever.base.request import dispatch_request
 from opengever.base.request import expect_ok_response
@@ -9,7 +10,6 @@ from opengever.base.response import IResponseContainer
 from opengever.base.types import UnicodeCoercingText
 from opengever.base.utils import escape_html
 from opengever.document.interfaces import ICheckinCheckoutManager
-from opengever.globalindex.model import WORKFLOW_STATE_LENGTH
 from opengever.meeting import _
 from opengever.meeting.activity.activities import ProposalDecideActivity
 from opengever.meeting.activity.activities import ProposalScheduledActivity
